Refactor react-test-renderer into native renderer
