+++ src/core/asyncUtils.js
@@ -1,12 +1,10 @@
-import { act } from 'react-test-renderer'
-
 function createTimeoutError(utilName, { timeout }) {
   const timeoutError = new Error(`Timed out in ${utilName} after ${timeout}ms.`)
   timeoutError.timeout = true
   return timeoutError
 }
 
-function asyncUtils(addResolver) {
+function asyncUtils(act, addResolver) {
   let nextUpdatePromise = null
 
   const waitForNextUpdate = async (options = {}) => {
