+++ src/core/cleanup.js
@@ -16,4 +16,13 @@ function removeCleanup(callback) {
   cleanupCallbacks = cleanupCallbacks.filter((cb) => cb !== callback)
 }
 
+cleanup.autoRegister = function () {
+  // Automatically registers cleanup in supported testing frameworks
+  if (typeof afterEach === 'function' && !process.env.RHTL_SKIP_AUTO_CLEANUP) {
+    afterEach(async () => {
+      await cleanup()
+    })
+  }
+}
+
 export { cleanup, addCleanup, removeCleanup }
