+++ src/native/index.js
@@ -1,0 +1,5 @@
+import { cleanup } from '../core'
+
+cleanup.autoRegister()
+
+export * from './pure'
